Fix Copy-paste error
Trivial Copy-Paste error.

Note: I could not verify that the parameter works at all… The website uses v4 of the API, there it works (with just the name of the company no hash needed). So it might be better to "upgrade" to v4 in general? But that would be a separate, bigger PR.